Don't try to validate {start,end}_time as times.
MDES defines times in the format hh:mm, but JSON Schema defines times as
hh:mm:ss.  We _could_ replace the time format with an appropriate regex,
but I don't think it's necessary: the only thing the field
application is doing with the received values is display.  (Maybe a
maximum length validation would be useful later on.)