A hack for the Ruby-based json-schema validator.
The validator does not properly handle additionalProperties: false in
conjunction with extends[0].  There is a patch[1], but the patch
implements the example, not the wording, of draft 03's description of
"extends": it expects extends to be a schema URI, not a schema in
itself.

My reading of Section 5.26 of Draft 03 of JSON Schema indicates to me
that defining an empty schema for the "version" attribute and extending
it with the real versioned entity schema should be equivalent to just
extending the versioned entity schema; therefore, in my opinion, this is
the easiest workaround until we can get a more correct JSON Schema
validator in place.  (That, or a corrected draft with accompanying
validator code shows up.)

[0] https://github.com/hoxworth/json-schema/issues/31
[1] http://git.io/0-ZapA